requirements.txt
sounddevice
SoundFile
pyaudio
alive_progress
pyrebase
psutil
py-cpuinfo
flask
uuid
numpy
flwr
Pyrebase
Pillow
seaborn
pandas
matplotlib
sklearn
plotly
tqdm
pathlib
sounddevice
ipython
pyaudio
sox
asttokens
atomicwrites
attrs
backcall
certifi
charset-normalizer
colorama
cycler
decorator
executing
fonttools
idna
iniconfig
jedi
kiwisolver
matplotlib-inline
packaging
parso
pickleshare
pip
pluggy
prompt-toolkit
pure-eval
py
Pygments
pyparsing
pytest
python-dateutil
pytz
requests
setuptools
six
stack-data
tomli
traitlets
typing_extensions
urllib3
wcwidth
matplotlib
ipython
sounddevice
soundfile
alive_progress
librosa
boto3
scipy
pil
librosa
